requirements.txt
aiocfscrape==1.0.0
aiohttp==3.9.3
aiohttp-proxy==0.1.2
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
better-proxy==1.1.5
colorama==0.4.6
DateTime==5.5
frozenlist==1.4.1
Js2Py==0.74
loguru==0.7.2
multidict==6.0.5
pyaes==1.6.1
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyjsparser==2.7.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
soupsieve==2.5
TgCrypto==1.2.5
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
websockets==12.0
win32-setctime==1.1.0
yarl==1.9.4
zope.interface==6.4.post2
aiofiles~=24.1.0
brotli==1.1.0
pillow
tenacity